Add in a basic thermal simulation
Simple 1D thermal simulation that provides updrafts/downdrafts for the
butterfly. This includes heating/cooling of the ground and air columns;
updraft/downdrafts are created when sections of the air are
significantly warmer or cooler than the rest of the air.

At the moment it provides some mild entertainment but needs improving to
make it anything vaguely playable...